➕ updates email front and back
plus, separated concerns with queries
